move subprocess tests
